Improve sensor mast control
The sonar sensor reading greatly depends on the angle between the sensor
and the obstacle surface. A shallow angle reduces the ability to detect
an obstacle. It seems like the material of the obstacle also has an
impact on this.

To circumvent this limitation, we increase the rotation range and speed
of the sensor mast.

In remote mode, we stop the mast rotation when an obstacle is detected.
This way the sensor will remain focused on the detected obstacle.